Moved methods and tests around
