+++ src/matrix/mod.rs
@@ -1,7 +1,7 @@
 use crate::{floating_point::FloatingPoint, tuple::Tuple};
-use axis::Axis;
 
 pub mod axis;
+pub mod transformation;
 
 #[cfg(test)]
 mod tests;
@@ -55,32 +55,6 @@ impl Matrix {
         }
 
         Matrix::new(size, size, Option::Some(elements))
-    }
-
-    pub fn translation(x: f64, y: f64, z: f64) -> Self {
-        let mut translation = Matrix::identity(4);
-        translation.elements[0][3] = x;
-        translation.elements[1][3] = y;
-        translation.elements[2][3] = z;
-
-        translation
-    }
-
-    pub fn scaling(x: f64, y: f64, z: f64) -> Self {
-        let mut scaling = Matrix::identity(4);
-        scaling.elements[0][0] = x;
-        scaling.elements[1][1] = y;
-        scaling.elements[2][2] = z;
-
-        scaling
-    }
-
-    pub fn reflect(axis: Axis) -> Self {
-        return match axis {
-            Axis::X => Matrix::scaling(-1.0, 1.0, 1.0),
-            Axis::Y => Matrix::scaling(1.0, -1.0, 1.0),
-            Axis::Z => Matrix::scaling(1.0, 1.0, -1.0)
-        };
     }
 
     pub fn multiply(&self, other: &Matrix) -> Matrix {
